Fix word break issues in list groups
Currently list groups have line breaks in the middle of words quite
frequently for me. This addresses the issue, and prioritises line breaks
at the ends of words rather than the middle of them.

If a word is longer than the container, then it will still break in the
middle of the word, but I think that's fine – it's better than it
overflowing.